% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_slider_page.R
\name{audio_slider_page}
\alias{audio_slider_page}
\title{Slider page to match to audio files}
\usage{
audio_slider_page(
  label,
  item_idx = "",
  ref_src = NULL,
  stm_base = NULL,
  sliderSounds = NULL,
  audioFormat_stimulus = NULL,
  ref_buttonSize = "20\%",
  stm_buttonSize = "20\%",
  min = NULL,
  max = NULL,
  step = 1,
  value,
  save_answer = TRUE,
  dict = raleR::RALE_dict,
  admin_ui = NULL
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{item_idx}{Item Index to display in header.}

\item{ref_src}{audio files for (static) reference sounds (character vector)}

\item{stm_base}{"body" for audio files that are mapped onto the slider. If used, files must be named "body\if{html}{\out{<Nr>}}.\if{html}{\out{<format>}}"
and \code{min} and \code{max} have to be defined. Mutually exclusive with \code{sliderSounds}.}

\item{sliderSounds}{character vector containing the audio files that should be mapped onto the slider. The slider is scaled from 0 to \code{length(sliderSounds)+1}. Mutually exclusive with \code{stm_base}.}

\item{audioFormat_stimulus}{file-format of the sounds mapped to the slider. Only required if \code{stm_base} is used rather than \code{sliderSounds} and file format differs from (first) specified reference sound.}

\item{ref_buttonSize}{character vector for the widths of each reference play button. Needs to contain a unit such as "px" or "cm".}

\item{stm_buttonSize}{character vector for the width of the button to play back the slider-manipulated sound. Needs to contain a unit such as "px" or "cm".}

\item{min, max}{minimum and maximum value for the slider. Only required if \code{stm_base} is used rather than \code{sliderSounds}}

\item{step}{slider resolution}

\item{value}{starting value for the slider.}

\item{save_answer}{Should the answer the be saved?}

\item{dict}{\code{i18n} dictionary for internationalisation.}

\item{admin_ui}{Optional UI for the admin interface. If \code{NULL}, no admin interface is created.}
}
\description{
Creates a page where the participant uses a slider to change one audio file to match another.
Both souunds can be played back with toggle buttons. The selected slider value is the stored response.
}
\note{
The RStudio preview function seems not to work
for slider pages on some machines.
}
