% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_raleUpdate.R
\name{DEG_raleUpdate}
\alias{DEG_raleUpdate}
\title{Update of the psyquest-DEG function}
\usage{
DEG_raleUpdate(subscales = c(), language)
}
\arguments{
\item{subscales}{A character vector of subscales to be included in the questionnaire.}

\item{language}{Language for testing.}
}
\description{
Wrapper for the \code{DEG} function of the \code{psyquest} package. Prior to calling the function,
some adjustments on the dictionary (\code{psyquest::psyquest_dict}) are made to make it more formal.
}
