% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_datapolicy.R
\name{intro_datapolicy}
\alias{intro_datapolicy}
\title{Introduction and Data Policy pages}
\usage{
intro_datapolicy(dict = raleR::RALE_dict)
}
\arguments{
\item{dict}{\code{i18n} dictionary to be used for internationalisation.}
}
\description{
Creates a welcome page and a data policy page. The texts have been approved by the DGM ethics committee and must not be changed. Only useful for this particular study.
}
