% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAudio_NAFC_page.R
\name{multiAudio_NAFC_page}
\alias{multiAudio_NAFC_page}
\title{NAFC page with multiple audio stimuli}
\usage{
multiAudio_NAFC_page(
  label,
  refAlt_src,
  cmpAlt_src = NULL,
  prompt = "Please listen to all sounds and select the correct answer.",
  refAlt_labels = LETTERS[1:length(refAlt_src)],
  cmpAlt_labels = paste0("X", 1:length(cmpAlt_src)),
  selectionQuestion = "Which sound is correct?",
  selectionLabels = refAlt_labels,
  selectionValues = 1:length(refAlt_src),
  correct_answer = NULL,
  defaultSelection = "",
  refAlt_buttonSize = "20\%",
  cmpAlt_buttonSize = refAlt_buttonSize,
  msg_validation = "Please select an answer!",
  inline = FALSE,
  save_answer = TRUE,
  is_final_page = FALSE,
  button_text = "Next"
)
}
\arguments{
\item{label}{page label}

\item{refAlt_src, cmpAlt_src}{URLs (character vector) of the files for reference sounds (upper row)
and comparison sounds (lower row)}

\item{prompt}{additional text above the buttons toggling the sounds}

\item{refAlt_labels, cmpAlt_labels}{Labels (character vector) for the buttons triggering the reference
sounds (upper row) and comparison sounds (lower row)}

\item{selectionQuestion}{Question above the answers (radio buttons).}

\item{selectionLabels}{answer options}

\item{selectionValues}{values returned by the radio buttons}

\item{correct_answer}{Correct choice option.}

\item{defaultSelection}{Default selection for the radio buttons. Default is an empty string, i.e. no selection.}

\item{refAlt_buttonSize, cmpAlt_buttonSize}{character: width of the button toggling the reference and
comparison sounds. Must contain the width itself and the unit, i.e. '\%' or 'px' and will be written
to the\code{style} attribute.}

\item{msg_validation}{message that is displayed if no answer is selected when the user clicks the "Next" button.}

\item{inline}{should radios button be aligned horizontically. Default is \code{FALSE}.}

\item{save_answer}{save answer to disk. Default is \code{TRUE}.}

\item{is_final_page}{specify if this is the last page of the test. Default is \code{FALSE}.}

\item{button_text}{text for button that navigates to next page}
}
\description{
NAFC page to compare different sound. Reference alternatives (upper row)
and comparison alternatives(lower row) can be independently created and require audio sources and button labels.
Radio buttons are used to select which reference or comparing alternative respectively meets a given
criterion
}
