% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{RALE}
\alias{RALE}
\alias{NoiSeQR}
\alias{EXP}
\title{Room Acoustic Listening Expertise Questionnaires}
\usage{
RALE(
  label = "RALE",
  subscales = c("TP", "SL", "RR", "L"),
  random_order = TRUE,
  dict = raleR::RALE_dict
)

NoiSeQR(
  label = "NOISEQR",
  subscales = c("WORK", "SLEEP", "HABITATION"),
  random_order = TRUE,
  dict = raleR::RALE_dict
)

EXP(label = "EXP", random_order = TRUE, dict = raleR::RALE_dict)
}
\arguments{
\item{label}{Label (\code{string}) for the saved results. Question identifiers will be added automatically.}

\item{subscales}{Subscales to choose from each inventory. For \code{RALE} the subscales \code{TP} (timbre perception), \code{SL} (sound localization),
\code{RR} (room reverberation) and \code{L} (loudness perception) are available. For \code{NoiSeQR} the \code{GENERAL} noise sensitivity and the subscales
\code{WORK}, \code{SLEEP}  and \code{HABITATION} are available. For \code{EXP}, there are no subscales.}

\item{random_order}{Logical value that defines whether the items should be presented in a random order. Default is \code{TRUE}.}

\item{dict}{\code{i18n} dictionary for internationalisation.}
}
\description{
\code{RALE}, \code{NoiSeQR} and \code{EXP} implement questionnaires to measure room acoustic listening expertise (RALE), noise sensitivity
using the reduced noise sensitivity questionnaire by Schuette al al. (NoiSeQR) and previous experience with room acoustics (EXP).
}
\references{
Schuette, M., Marks, A., Wenning E., & Griefahn, B. (2007). The development of the noise sensitivity questionnaire. Noise and Health 9(23), pp.15-24. DOI: 10.4103/1463-1741.34700
}
