% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RALE_dict}
\alias{RALE_dict}
\title{RALE dictionary}
\format{
An object of class \code{i18n_dict} (inherits from \code{R6}) of length 8.
}
\usage{
RALE_dict
}
\description{
A dictionary for internationalisation of all tests and questionnaires. Currently available languages are English (en) and German (de, de_f).
So far the wording does not differ between formal and standard German.
}
\keyword{datasets}
