% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RALE_education.R
\name{RALE_education}
\alias{RALE_education}
\title{Inquire educational level}
\usage{
RALE_education(label = "education", dict = raleR::RALE_dict)
}
\arguments{
\item{label}{label for storing results (character scalar).}

\item{dict}{\code{i18n} dictionary for internationalisation.}
}
\description{
Multiple choice question for educational degree with seven response options
}
