% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_pick_TPT_RMT.R
\name{rand_pick_TPT_RMT}
\alias{rand_pick_TPT_RMT}
\title{Randomly choose either the TPT or RMT}
\usage{
rand_pick_TPT_RMT(
  TPT_with_welcome = TRUE,
  TPT_with_training = TRUE,
  RMT_url_dir,
  RMT_src = "Git2_",
  RMT_N_items = NULL,
  RMT_targets = NULL,
  RMT_min = 0,
  RMT_max = 50,
  RMT_sliderLength = 20
)
}
\arguments{
\item{TPT_with_welcome}{logical value indicating whether the TPT should include a welcome page.}

\item{TPT_with_training}{logical value indicating whether the TPT should include a training phase.}

\item{RMT_url_dir, RMT_src, RMT_N_items, RMT_targets, RMT_min, RMT_max, RMT_sliderLength}{RMT configuration parameters that are passed on to \link{RMT_battery}.}
}
\description{
Creates either a TPT or an RMT instance. The selection is random.
}
