% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMT_battery.R
\name{RMT_battery}
\alias{RMT_battery}
\title{RMT item battery}
\usage{
RMT_battery(
  N_items = NULL,
  label = "RMT",
  pool_sounds = NULL,
  stm_base = NULL,
  min = NULL,
  max = NULL,
  sliderLength = length(pool_sounds),
  targets = NULL,
  target_range = NULL,
  target_margins = 3,
  show_trial_number = TRUE,
  format_numbers = "\%02d",
  file_ext = "mp3",
  dict = raleR::RALE_dict,
  ...
)
}
\arguments{
\item{N_items}{number of items in the test battery. If \code{targets} is provided, this argument is ignored.}

\item{label}{label for the items in the test battery. Item numbers will be added automatically}

\item{pool_sounds}{character vector with audio files that can be used in the test battery. If \code{NULL}, the function will
create a pool of sounds based on \code{stm_base}, \code{min}, \code{max} and \code{format_numbers}.}

\item{stm_base}{"body" for audio files that are mapped onto the slider. If used, files must be named "body\if{html}{\out{<Nr>}}.\code{file_ext}"
and \code{min} and \code{max} have to be defined. Mutually exclusive with \code{pool_sounds}.}

\item{min}{minimum value for the slider. Only required if \code{stm_base} is used rather than \code{pool_sounds}.}

\item{max}{maximum value for the slider. Only required if \code{stm_base} is used rather than \code{pool_sounds}.}

\item{sliderLength}{length of the slider, i.e. number of sounds that can be selected. Default is the length of
\code{pool_sounds} so that each item presents the full scale.}

\item{targets}{numeric vector with indices of sounds that should be used as targets in the test battery. If \code{NULL},
the function will sample \code{N_items} targets from \code{pool_sounds}.}

\item{target_range}{numeric vector with indices of sounds that can be used as targets in the test battery. If \code{NULL},
the function will sample from the full range of \code{pool_sounds}.}

\item{target_margins}{numeric value that defines the margins around the target that should be excluded from the sampled range.}

\item{show_trial_number}{logical value that defines whether the trial number should be displayed in the prompt header.}

\item{format_numbers}{format string for adding numbers to the base URL when building \code{pool_sounds}
(e.g. "\%02d" for 2 digits integers).}

\item{file_ext}{file extension for the audio files in \code{pool_sounds}. Default is "mp3".}

\item{dict}{dictionary object used for internationalisation.}

\item{...}{additional arguments that are passed to \code{\link{audio_slider_page}()}.}
}
\description{
Create a battery of items for the Reverberation Matching Test (RMT), including instructions, item battery,
scoring and feedback. Expects a web directory with audio files that are named after the pattern "\code{source}\if{html}{\out{<Nr>}}_\if{html}{\out{<itemNr>}}.\if{html}{\out{<format>}}".
Targets for the items are either provided manually or generated randomily within specified margins ensuring that the target is never the end point of a slider range.
This function returns a \code{psychTestR::module()} object that can be used in a \code{psychTestR} test.
}
