% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRT_abx_page.R
\name{RRT_abx_page}
\alias{RRT_abx_page}
\title{RRT item page with ABX interface}
\usage{
RRT_abx_page(
  url_A,
  url_B,
  label,
  prompt,
  question,
  labels_AB = c("Play A", "Play B"),
  label_X = "Play X",
  selectionLabels = c("A = X", "B = X"),
  selectionValues = c("A", "B"),
  inline = TRUE,
  fixed_X = NULL,
  dict = raleR::RALE_dict,
  ...
)
}
\arguments{
\item{url_A}{character or vector (for interchangeable sources) of URLs for A}

\item{url_B}{character or vector (for interchangeable sources) of URLs for B}

\item{label}{page label}

\item{prompt}{additional text above the buttons toggling the sounds}

\item{question}{Question above the answers (radio buttons).}

\item{labels_AB}{Labels (character vector) for the buttons triggering the sound A and B
sounds (upper row)}

\item{label_X}{Label (character vector for the button) for triggering the sound X}

\item{selectionLabels}{answer options}

\item{selectionValues}{values returned by the radio buttons}

\item{inline}{should answer option be aligned horizontally? Default is \code{TRUE}.}

\item{fixed_X}{If desired give a source that will always be set for X. Default is \code{NULL}, i.e. all sources will be randomly assigned to A, B, and X.}

\item{dict}{\code{i18n} dictionary for internationalisation}

\item{...}{other parameters to be passed on to \code{\link{multiAudio_NAFC_page}()}.}
}
\description{
ABX page to with two reference sound and an comparison sound that matches either A or B.
Radio buttons are used to select whether A or B matches X. The function is a wrapper for the multiAudio_NAFC_page.
}
