% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRT_battery.R
\name{RRT_battery}
\alias{RRT_battery}
\title{RRT item battery}
\usage{
RRT_battery(
  N_items = 20,
  url_dir,
  source = "Git",
  N_source = 3,
  baseline = 0,
  range = 0:N_items,
  audio_file_type = "mp3",
  label = "RRT",
  labels_AB = c("Play A", "Play B"),
  label_X = "Play X",
  dict = raleR::RALE_dict,
  ...
)
}
\arguments{
\item{N_items}{number of items in the test battery. If \code{targets} is provided, this argument is ignored.}

\item{url_dir}{URL to the directory with the audio files.}

\item{source}{source Identifier in the audio file names. Should include everything in the file name except for the source and stimulus number.}

\item{N_source}{number of different source signals. If 3, all sounds A, B and X have different source signals. If 1, all have the same.
If 2, A and B have the same source and X is different.}

\item{baseline}{Which stimulus item number should be equal in all items?}

\item{range}{numeric vector with item numbers that can be used in the test battery. Length must be larger than or equal to \code{N_items}.}

\item{audio_file_type}{file extension for the audio files in \code{url_dir}.}

\item{label}{Label for the items in the test battery. Item numbers will be added automatically.}

\item{labels_AB}{Labels for the buttons toggling the audio playback for sounds A and B.}

\item{label_X}{Label for the button toggling the audio playback for sound X.}

\item{dict}{\code{i18n} dictionary used for interationalisation.}

\item{...}{Other arguments passed on to \code{\link{RRT_abx_page}()}.}
}
\description{
Create a battery of items for the Room Recognition Test (RRT), including instructions, item battery, scoring and feedback.
Within the battery, \code{N_items} are randomly drawn from the supplied \code{range}. Items are administered in a random order.
Expects a web directory with audio files that are named after the pattern "\code{source}\if{html}{\out{<Nr>}}_\if{html}{\out{<itemNr>}}.\if{html}{\out{<format>}}".
This function returns a \code{psychTestR::module()} object that can be used in a \code{psychTestR} test.
}
