% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLevel.R
\name{setLevel}
\alias{setLevel}
\title{Page to set the playback level}
\usage{
setLevel(url_dir, filename, dict = raleR::RALE_dict)
}
\arguments{
\item{url_dir}{The URL directory where the test audio file is stored.}

\item{filename}{The name of the test audio file within the directory (including file format extension such as 'wav').}

\item{dict}{\code{i18n} dictionary for internationalisation.}
}
\description{
Page for playback level adjustment. A sound of long noise bursts at -51 dBFS and short noise bursts at -57dBFS is presented.
Participants are instructed to adjust the playback level so that the long noise bursts are barely audible and the short ones are not.
The test sound can be triggered with a toggle button.
}
