% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_block_intro.R
\name{test_block_intro}
\alias{test_block_intro}
\title{Intro pages for test blocks}
\usage{
test_block_intro(test_key, dict = raleR::RALE_dict)
}
\arguments{
\item{test_key}{Identifiert for the test block in the \code{dict} object.}

\item{dict}{\code{i18n} dictionary used for internationalisation.}
}
\description{
creates a one button page to display introduction text for a new test block.
}
